MIDI出力対応と再生ロジック分離、編集UI文言/アイコン改善 ### 概要 以下をまとめて実施しました。
- 保存セクションに `MIDI出力` ボタンを追加
- 再生イベント変換・MIDI生成ロジックを `main.ts` から `playback.ts` に分離
- 編集アクションの文言を明確化し、各ボタンにSVGアイコンを追加
- 操作ボタンの並びと視認性を改善（休符を音符化 / 音符分割 / 音符削除）

### 主な変更点

1. MIDI出力機能の追加
- `mikuscore-src.html` に `MIDI出力` ボタン（`#downloadMidiBtn`）を追加
- `src/ts/main.ts` に `onDownloadMidi()` を追加し、`.mid` ダウンロードを実装
- 保存成功時のみ MusicXML/MIDI 出力ボタンを有効化

2. 再生処理のモジュール分離
- `src/ts/playback.ts` を新規追加
  - `buildPlaybackEventsFromXml(xml, ticksPerQuarter)`
  - `buildMidiBytesForPlayback(events, tempo)`
- `main.ts` から再生イベント生成・MIDI生成関連の実装を削除し、`playback.ts` を import する構成へ整理

3. MIDI生成品質の改善
- イベント配置を `startTick` 指定ベースに対応（時間位置の明示）
- チャンネルごとの `ProgramChangeEvent` を追加し、GM `Electric Piano 2 (instrument: 5)` を指定（ch10除外）

4. 編集UI改善
- 文言変更
  - `編集確定` → `小節編集確定`
  - `編集破棄` → `小節編集破棄`
  - `小節を再生` → `小節再生`
- SVGアイコン付きボタンへ更新
- 休符を音符化ボタンのアイコン視認性を改善（横長アイコン対応）

### 変更ファイル
- `mikuscore-src.html`
- `mikuscore.html`
- `src/css/app.css`
- `src/js/main.js`
- `src/ts/main.ts`
- `src/ts/playback.ts`（新規）

### 期待効果
- 出力形式の拡張（MusicXML + MIDI）
- `main.ts` の責務分離による保守性向上
- 編集操作の意味が伝わりやすいUIへ改善
- MIDI側のタイミング/音色指定の再現性向上
