+++ mikuscore-src.html
@@ -120,9 +120,25 @@
             <div id="measureEditorArea" class="ms-debug-score"></div>
           </div>
           <div class="ms-actions">
-            <button id="measureApplyBtn" type="button" class="md-button md-button--primary" disabled>編集確定</button>
-            <button id="measureDiscardBtn" type="button" class="md-button md-button--surface" disabled>編集破棄</button>
-            <button id="playMeasureBtn" type="button" class="md-button md-button--tonal" disabled>小節を再生</button>
+            <button id="measureApplyBtn" type="button" class="md-button md-button--primary ms-icon-button" disabled>
+              <svg aria-hidden="true" viewBox="0 0 24 24" class="ms-btn-icon" fill="none" stroke="currentColor" stroke-width="1.9" stroke-linecap="round" stroke-linejoin="round">
+                <path d="M20 7L9 18l-5-5"></path>
+              </svg>
+              <span>小節編集確定</span>
+            </button>
+            <button id="measureDiscardBtn" type="button" class="md-button md-button--surface ms-icon-button" disabled>
+              <svg aria-hidden="true" viewBox="0 0 24 24" class="ms-btn-icon" fill="none" stroke="currentColor" stroke-width="1.9" stroke-linecap="round" stroke-linejoin="round">
+                <path d="M18 6L6 18"></path>
+                <path d="M6 6l12 12"></path>
+              </svg>
+              <span>小節編集破棄</span>
+            </button>
+            <button id="playMeasureBtn" type="button" class="md-button md-button--tonal ms-icon-button" disabled>
+              <svg aria-hidden="true" viewBox="0 0 24 24" class="ms-btn-icon" fill="currentColor">
+                <path d="M8 6.5v11l9-5.5z"></path>
+              </svg>
+              <span>小節再生</span>
+            </button>
           </div>
 
           <div class="ms-grid">
@@ -154,9 +170,41 @@
           </div>
 
           <div class="ms-actions">
-            <button id="splitNoteBtn" type="button" class="md-button md-button--tonal">音符分割</button>
-            <button id="convertRestBtn" type="button" class="md-button md-button--tonal">休符を音符化</button>
-            <button id="deleteBtn" type="button" class="md-button md-button--surface">音符削除</button>
+            <button id="convertRestBtn" type="button" class="md-button md-button--tonal ms-icon-button">
+              <svg aria-hidden="true" viewBox="0 0 34 24" class="ms-btn-icon ms-btn-icon--wide" fill="none" stroke="currentColor" stroke-width="2.2" stroke-linecap="round" stroke-linejoin="round">
+                <!-- quarter rest (wide, readable silhouette) -->
+                <path d="M6.5 3.6c4 1.7 5.4 4.2 2.9 7.3-1.2 1.4-1.2 2.7.6 4.1 1.5 1.2 3.6 2 5.2 3.5"></path>
+                <path d="M9.2 7.6c2.2 1.2 3.4 2.8 2 4.6"></path>
+                <path d="M11.8 15.3c1.6.8 2.8 1.5 4 2.6"></path>
+                <!-- convert arrow -->
+                <path d="M16.8 12h5.2"></path>
+                <path d="M20.2 9.5l2.6 2.5-2.6 2.5"></path>
+                <!-- quarter note -->
+                <circle cx="27.3" cy="15.9" r="2.2" fill="currentColor" stroke="none"></circle>
+                <path d="M29.3 15.9V7.1"></path>
+              </svg>
+              <span>休符を音符化</span>
+            </button>
+            <button id="splitNoteBtn" type="button" class="md-button md-button--tonal ms-icon-button">
+              <svg aria-hidden="true" viewBox="0 0 24 24" class="ms-btn-icon" fill="none" stroke="currentColor" stroke-width="1.9" stroke-linecap="round" stroke-linejoin="round">
+                <circle cx="7.5" cy="15.5" r="2.5" fill="currentColor" stroke="none"></circle>
+                <path d="M10 15.5V8"></path>
+                <path d="M12 6v12"></path>
+                <circle cx="16.5" cy="15.5" r="2.5" fill="currentColor" stroke="none"></circle>
+                <path d="M19 15.5V8"></path>
+              </svg>
+              <span>音符分割</span>
+            </button>
+            <button id="deleteBtn" type="button" class="md-button md-button--surface ms-icon-button">
+              <svg aria-hidden="true" viewBox="0 0 24 24" class="ms-btn-icon" fill="none" stroke="currentColor" stroke-width="1.9" stroke-linecap="round" stroke-linejoin="round">
+                <path d="M4 7h16"></path>
+                <path d="M9 7V5h6v2"></path>
+                <rect x="6.5" y="7" width="11" height="13" rx="1.5"></rect>
+                <path d="M10 10v7"></path>
+                <path d="M14 10v7"></path>
+              </svg>
+              <span>音符削除</span>
+            </button>
           </div>
         </section>
 
@@ -164,6 +212,7 @@
           <h2 class="md-section-title"><span class="ms-step-no">4</span>保存</h2>
           <div class="ms-actions">
             <button id="downloadBtn" type="button" class="md-button md-button--secondary" disabled>MusicXML出力</button>
+            <button id="downloadMidiBtn" type="button" class="md-button md-button--tonal" disabled>MIDI出力</button>
           </div>
         </section>
       </div>
